Project-social-media/app/frontend/requirements.txt
streamlit==1.37.0
requests==2.32.3
pillow==10.2.0